+++ Assets/Scripts/Ball/DisturbBall.cs
@@ -1,7 +1,6 @@
 using System;
 using R3;
 using UnityEngine;
-using SafeEventSystem;
 
 public class DisturbBall : BallBase
 {
@@ -10,7 +9,7 @@ public class DisturbBall : BallBase
     {
         base.InitBall(d, rank, level);
         this.isMergable = false;
-        _disposable = SafeEventManager.OnBallMergedSimple.Subscribe(CheckNearMerge).AddTo(this);
+        _disposable = EventManager.OnBallMergedSimple.Subscribe(CheckNearMerge).AddTo(this);
     }
     
     private void CheckNearMerge((BallBase ball1, BallBase ball2) mergeData)
