+++ Assets/Scripts/Merge/MergeManager.cs
@@ -6,7 +6,6 @@ using DG.Tweening;
 using TMPro;
 using UnityEngine.EventSystems;
 using unityroom.Api;
-using SafeEventSystem;
 
 public class MergeManager : MonoBehaviour
 {
@@ -230,7 +229,7 @@ public class MergeManager : MonoBehaviour
         
         // SafeEventManagerを適用
         var attackData = AttackData.FromDictionary(_attackCounts);
-        attackData = SafeEventManager.OnPlayerAttack.Process(attackData);
+        attackData = EventManager.OnPlayerAttack.Process(attackData);
         _attackCounts = attackData.ToDictionary();
         
         GameManager.Instance.EnemyContainer.AttackEnemy(type, (int)atk).Forget();
@@ -284,13 +283,13 @@ public class MergeManager : MonoBehaviour
         CurrentBall.transform.SetParent(_ballContainer.transform);
         
         // ボールドロップイベントを発火
-        SafeEventManager.OnBallDropSimple.OnNext(R3.Unit.Default);
+        EventManager.OnBallDropSimple.OnNext(R3.Unit.Default);
     }
 
     private void SkipBall()
     {
         // ボールスキップイベントを発火
-        SafeEventManager.OnBallSkip.OnNext(R3.Unit.Default);
+        EventManager.OnBallSkip.OnNext(R3.Unit.Default);
         Destroy(CurrentBall);
     }
