+++ Assets/Scripts/Relic/CreateBallWhenStatusEffect.cs
@@ -1,6 +1,5 @@
 using UnityEngine;
 using R3;
-using SafeEventSystem;
 
 /// <summary>
 /// プレイヤーにステータス効果が追加されたときにランダムボールを生成するレリック
@@ -10,7 +9,7 @@ public class CreateBallWhenStatusEffect : RelicBase
     protected override void RegisterEffects()
     {
         // プレイヤーステータス効果追加時のイベント購読
-        var subscription = SafeEventManager.OnPlayerStatusEffectAdded.Subscribe(OnStatusEffectAdded);
+        var subscription = EventManager.OnPlayerStatusEffectAdded.Subscribe(OnStatusEffectAdded);
         _simpleSubscriptions.Add(subscription);
     }
