+++ Assets/Scripts/Relic/Fukiya.cs
@@ -1,6 +1,5 @@
 using UnityEngine;
 using R3;
-using SafeEventSystem;
 
 /// <summary>
 /// オーガナイズ時にプレイヤーに10ダメージを与えるレリック
@@ -10,7 +9,7 @@ public class Fukiya : RelicBase
     protected override void RegisterEffects()
     {
         // オーガナイズ時のイベント購読
-        var subscription = SafeEventManager.OnOrganise.Subscribe(OnOrganise);
+        var subscription = EventManager.OnOrganise.Subscribe(OnOrganise);
         _simpleSubscriptions.Add(subscription);
     }
