--- Assets/Scripts/System/AttackDataModifier.cs
@@ -1,6 +1,0 @@
-// このファイルは削除されました
-// AttackDataModifierは新しいSafeEventSystemのValueProcessorsで置き換えられました
-// 使用方法:
-// - ValueProcessors.AddAttack(AttackType.Normal, 5)
-// - ValueProcessors.MultiplyAttack(2.0f)
-// - ValueProcessors.ConvertAttackType(AttackType.Normal, AttackType.All)
