--- Assets/Scripts/System/AttackDataModifier.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: af18b9d7185904ce0b392cbf2d63350b
